app, store: support attribute selectors
the 'attributes' projection is needed, but unfortunately
it changes the result format, so extra parsing logic
is added.

also, more type assertions are needed -
the results here always come back as an array, even if they're
simple ids, etc. - not attributes. this is not the case with
any other query.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>