client/inventory: get device(s) by ID
use the search endpoint to pull the reindexed device
from inventory.

it's a simple query by ID(s) - we want a single device for now, but
bulk GET is available.

it may be important for the 'final' implementation of the reindex EP,
which can bulk update/cache/etc. the triggers and data.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>