MEN-4970: provision and use a single elasticsearch index for devices
Update the service to provision and use a single, configurable
elasticsearch index to store the devices. The number of shards and
replicas is configurable. The defaults are one single shard, and no
replicas. To improve performances, the tenant ID is used as routing key
both when indexing and querying the data. In case of open-source, empty
routing key is used, which means the internal implementation in
elasticsearch will use the document ID.

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>